refac: move all files regarding edge cases for both reference and operational data in casedb to one folder.
